Used of an Index File for Library Loadup and Duplication Check.
- Large Speedup for startup times and resuming of blogs since we now
catalogize all downloaded filenames together with their URLs in a small
single index file, instead of checking for all single downloaded image
files in the download folder at startup and which is now also used for
duplication check. This should improve speed drastically. Also, its now
possible to safely remove images out of the
\Blogs\"MyDownloadedTumblrBlogFolder"\ without rendering in download
them again, as long as you keep the .tumblr (index) file in the Index
folder. This opens the way to a backup function.
- Specifying the number of Posts in each blogs. Might be equivalent to
number of pictures, if the blog only contains pictures.