Use the filename for duplicate check.
Some images were reloaded even if they where already downloaded since we
have saved the full url and checked for duplicates using it. If the file
however was hosted on a different mirror, the tool would redownload the
same file and increase the counter for downloaded images even if it was
already downloaded.